refactor: group conversion interfaces
