+++ backend/src/DataModeling/Converter/Csharp/JsonMetadataToCsharpConverter.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Altinn.Studio.DataModeling.Converter.Interfaces;
 using Altinn.Studio.DataModeling.Metamodel;
 using static Altinn.Studio.DataModeling.Utils.RestrictionsHelper;
